infra/functions/budget_guard/requirements.txt
google-api-python-client>=2.0.0,<3.0.0
google-auth>=2.0.0,<3.0.0
